0.5.6
This release sees some added checks to make certain that people don't leave the DCS data folder empty when trying to start a server. If it is it will not start.

It also adds an option to turn on/off the forced 24 hour server restarts.

And the option to turn on and off the monitor for GUI non-responsive state hangs in DCS. Along with a warning that DCS 2.7.0.xxxx currently doesn't seem to support this method of monitoring.